Before:
```
➜  workerpool git:(main) ✗ go test -bench . -benchtime=1000x -benchmem
goos: darwin
goarch: amd64
pkg: github.com/zaidsasa/workerpool
cpu: Intel(R) Core(TM) i7-4870HQ CPU @ 2.50GHz
BenchmarkWorkerPool-8               1000             54691 ns/op            2503 B/op        100 allocs/op
PASS
ok      github.com/zaidsasa/workerpool  2.256s
```

After: 
```
➜  workerpool git:(reduce-memory-alloc) go test -bench . -benchtime=1000x -benchmem
goos: darwin
goarch: amd64
pkg: github.com/zaidsasa/workerpool
cpu: Intel(R) Core(TM) i7-4870HQ CPU @ 2.50GHz
BenchmarkWorkerPool-8               1000             24676 ns/op               9 B/op          0 allocs/op
PASS
ok      github.com/zaidsasa/workerpool  2.232s
```